n! (正確版)

+++ src/Main.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner scn = new Scanner(System.in);
-        long n = scn.nextInt();
+        int n = scn.nextInt();
         long a = 1;
-        for (long i = 1; i<=n ; i++ ){
+        for (int i = 1; i<=n ; i++ ){
             a = a*i;
         }
         System.out.println(a);
